remove scheme mapping move blocking_operator to operator
