--- bindings/ocaml/src/block_operator.mli
@@ -1,18 +1,0 @@
-(* Generated by ocaml-rs *)
-
-open! Bigarray
-
-(* file: mod.rs *)
-
-type blocking_operator
-external new_blocking_operator_str: string -> (string * string) list -> (blocking_operator, string) Result.t  = "new_blocking_operator_str"
-external new_blocking_operator: Scheme.scheme -> (string * string) list -> (blocking_operator, string) Result.t  = "new_blocking_operator"
-external blocking_is_exist: blocking_operator -> string -> (bool, string) Result.t  = "blocking_is_exist"
-external blocking_create_dir: blocking_operator -> string -> (bool, string) Result.t  = "blocking_create_dir"
-external blocking_read: blocking_operator -> string -> (char array, string) Result.t  = "blocking_read"
-external blocking_write: blocking_operator -> string -> bytes -> (unit, string) Result.t  = "blocking_write"
-external blocking_copy: blocking_operator -> string -> string -> (unit, string) Result.t  = "blocking_copy"
-external blocking_rename: blocking_operator -> string -> string -> (unit, string) Result.t  = "blocking_rename"
-external blocking_delete: blocking_operator -> string -> (unit, string) Result.t  = "blocking_delete"
-external blocking_remove: blocking_operator -> string array -> (unit, string) Result.t  = "blocking_remove"
-external blocking_remove_all: blocking_operator -> string -> (unit, string) Result.t  = "blocking_remove_all"
