+++ bindings/ocaml/src/operator.mli
@@ -1,0 +1,17 @@
+(* Generated by ocaml-rs *)
+
+open! Bigarray
+
+(* file: mod.rs *)
+
+type operator
+external operator: string -> (string * string) list -> (operator, string) Result.t  = "operator"
+external blocking_is_exist: operator -> string -> (bool, string) Result.t  = "blocking_is_exist"
+external blocking_create_dir: operator -> string -> (bool, string) Result.t  = "blocking_create_dir"
+external blocking_read: operator -> string -> (char array, string) Result.t  = "blocking_read"
+external blocking_write: operator -> string -> bytes -> (unit, string) Result.t  = "blocking_write"
+external blocking_copy: operator -> string -> string -> (unit, string) Result.t  = "blocking_copy"
+external blocking_rename: operator -> string -> string -> (unit, string) Result.t  = "blocking_rename"
+external blocking_delete: operator -> string -> (unit, string) Result.t  = "blocking_delete"
+external blocking_remove: operator -> string array -> (unit, string) Result.t  = "blocking_remove"
+external blocking_remove_all: operator -> string -> (unit, string) Result.t  = "blocking_remove_all"
