--- bindings/ocaml/src/scheme.mli
@@ -1,79 +1,0 @@
-(* Generated by ocaml-rs *)
-
-open! Bigarray
-
-(* file: mod.rs *)
-
-type scheme = 
-| Azblob  (** [Azblob]: Azure Storage Blob services.*)
-
-| Azdfs  (** [Azdfs]: Azure Data Lake Storage Gen2. *)
-
-| Cacache  (** [Cacache]: cacache backend support. *)
-
-| Cos  (** [Cos]: Tencent Cloud Object Storage services. *)
-
-| Dashmap  (** [Dashmap]: dashmap backend support. *)
-
-| Etcd  (** [Etcd]: Etcd Services *)
-
-| Fs  (** [Fs]: POSIX alike file system. *)
-
-| Ftp  (** [Ftp]: FTP backend. *)
-
-| Gcs  (** [Gcs]: Google Cloud Storage backend. *)
-
-| Ghac  (** [Ghac]: GitHub Action Cache services. *)
-
-| Hdfs  (** [Hdfs]: Hadoop Distributed File System. *)
-
-| Http  (** [Http]: HTTP backend. *)
-
-| Ipfs  (** [Ipfs]: IPFS HTTP Gateway *)
-
-| Ipmfs  (** [Ipmfs]: IPFS mutable file system *)
-
-| Memcached  (** [Memcached]: Memcached service support. *)
-
-| Memory  (** [Memory]: In memory backend support. *)
-
-| MiniMoka  (** [MiniMoka]: Mini Moka backend support. *)
-
-| Moka  (** [Moka]: moka backend support. *)
-
-| Obs  (** [Obs]: Huawei Cloud OBS services. *)
-
-| Onedrive  (** [Onedrive]: Microsoft OneDrive services. *)
-
-| Gdrive  (** [Gdrive]: GoogleDrive services. *)
-
-| Dropbox  (** [Dropbox]: Dropbox services. *)
-
-| Oss  (** [Oss]: Aliyun Object Storage Services *)
-
-| Persy  (** [Persy]: persy backend support. *)
-
-| Redis  (** [Redis]: Redis services *)
-
-| Rocksdb  (** [Rocksdb]: RocksDB services *)
-
-| S3  (** [S3]: AWS S3 alike services. *)
-
-| Sftp  (** [Sftp]: SFTP services *)
-
-| Sled  (** [Sled]: Sled services *)
-
-| Supabase  (** [Supabase]: Supabase storage service *)
-
-| VercelArtifacts
-    (** [VercelArtifacts]: Vercel Artifacts service, as known as Vercel Remote Caching. *)
-
-| Wasabi  (** [Wasabi]: Wasabi service *)
-
-| Webdav  (** [Webdav]: WebDAV support. *)
-
-| Webhdfs  (** [Webhdfs]: WebHDFS RESTful API Services *)
-
-| Redb  (** [Redb]: Redb Services *)
-
-| Tikv  (** [Tikv]: Tikv Services *)
